giant refactoring: allow gpio line names and use upstream libgpiod bindings
This started from a desire to use a gpio line name (as set in device
tree) rather than a (chip, line offset) specification, since we've
seen the latter cause real problems when sometimes the pi-hat's
gpiochip became gpiochip2 instead of gpiochip1. Unfortunately, that
required somewhat invasive changes, since the code was built around
interfacing with a gpiochip instance. Also, the drivers/libgpiod CDLL
module didn't have the required wrapper for gpiod_find_line(const
char*).

So this refactors the code to use the Python binding module shipped
with upstream libgpiod, so all we need to remember for each relay is
an gpiod.Line instance, and updates the config file parsing to accept
a name instead of a chip,line pair. While there, it occured to me that
the "unique id per relay" could more easily be enforced by having
relays be a dict of dicts, instead of a list of dicts each having an
"id" key. The "don't specify the same gpio line twice" is
automatically enforced since the second time we'd request it we'd get
an error.

Also remove a few things that aiohttp have deprecated (the loop
argument and returning rather than raising a HTTPException).

I've tried to preserve the REST API, so other code interfacing with
the powerrelay service should not be affected. The config file format
is slightly changed, but that's by design: It forces people to update
their config, giving them an incentive to start using gpio line names.