main.py: Rewrite functionality to work with libgpiod 2.0
In libgpiod 1.x, requests were made on lines. So;
- get gpiochip
- line = request line X on gpiochip
- line.get_value() or line.set_value(<value>)

That have changed in gpiolib 2.0.
Now requests are made on the gpiochip (instead of on a line), and values
or set and get on the requests.
So now we:
- locate correct gpiochip
- request = request(gpiochip, [offset]) // offset can be multiple values
- request.set_values([ dict(<offset>, <values>) ] OR
    request.get_values[ <offset> ]

We cant just have a single reqest object, since we can have cases where
the gpios used are on different gpiochips, and the requests are bound
to a specific gpiochip. So we need to store multiple requests, so we can
use the requests to set/get value.

To be able to do that, and keep the same structure for setting/getting
the relays; create a new object RelayLine. Behaves somewhat the same
as when using gpiolib 1.x, meaning that we got an object that is tied
to a specific line.
We got all RelayLine's stored in the dict() lines variables, where
the key is the value specified in the configuration file.